feat: Reorganize EmulatorPass module and place gate-based device passes into  gate_device_passes module
