+++ src/braket/emulators/emulator_passes/__init__.py
@@ -1,8 +1,2 @@
-from braket.emulators.emulator_passes.criteria import (  # noqa: F401
-    ConnectivityCriterion,
-    EmulatorCriterion,
-    GateConnectivityCriterion,
-    GateCriterion,
-    QubitCountCriterion,
-)
+from braket.emulators.emulator_passes.emulator_criterion import EmulatorCriterion  # noqa: F401
 from braket.emulators.emulator_passes.emulator_pass import EmulatorPass, ProgramType  # noqa: F401
