--- src/braket/emulators/emulator_passes/criteria/__init__.py
@@ -1,13 +1,0 @@
-from braket.emulators.emulator_passes.criteria.connectivity_criterion import (  # noqa: F401
-    ConnectivityCriterion,
-)
-from braket.emulators.emulator_passes.criteria.emulator_criterion import (  # noqa: F401
-    EmulatorCriterion,
-)
-from braket.emulators.emulator_passes.criteria.gate_connectivity_criterion import (  # noqa: F401
-    GateConnectivityCriterion,
-)
-from braket.emulators.emulator_passes.criteria.gate_criterion import GateCriterion  # noqa: F401
-from braket.emulators.emulator_passes.criteria.qubit_count_criterion import (  # noqa: F401
-    QubitCountCriterion,
-)
