+++ src/braket/emulators/emulator_passes/gate_device_criteria/__init__.py
@@ -1,0 +1,12 @@
+from braket.emulators.emulator_passes.gate_device_criteria.connectivity_criterion import (  # noqa: F401 E501
+    ConnectivityCriterion,
+)
+from braket.emulators.emulator_passes.gate_device_criteria.gate_connectivity_criterion import (  # noqa: F401 E501
+    GateConnectivityCriterion,
+)
+from braket.emulators.emulator_passes.gate_device_criteria.gate_criterion import (  # noqa: F401 E501
+    GateCriterion,
+)
+from braket.emulators.emulator_passes.gate_device_criteria.qubit_count_criterion import (  # noqa: F401 E501
+    QubitCountCriterion,
+)
