+++ src/braket/emulators/emulator_passes/gate_device_criteria/gate_connectivity_criterion.py
@@ -6,7 +6,7 @@ from networkx.utils import graphs_equal
 from braket.circuits.circuit import Circuit
 from braket.circuits.compiler_directives import EndVerbatimBox, StartVerbatimBox
 from braket.circuits.gate import Gate
-from braket.emulators.emulator_passes.criteria.emulator_criterion import EmulatorCriterion
+from braket.emulators.emulator_passes import EmulatorCriterion
 from braket.registers.qubit_set import QubitSet
